logging

+++ loader.py
@@ -1,3 +1,4 @@
+import logging
 from typing import List
 import streamlink
 import subprocess
@@ -11,6 +12,9 @@ from db import db_session
 from db.models import *
 import time
 import sys
+from contextlib import redirect_stdout
+from io import StringIO
+import logging
 
 class StreamListener:
     oauth = '0gn793kk3c98a6ugz38tt7zgoq6i0g'
@@ -31,10 +35,12 @@ class StreamListener:
 
     def run(self):
         """начальная инициализация"""
+
         self.engine, self.session_maker = db_session.get_sessionmaker()
 
         self.phrazes = self.load_phrazes()
-        print('starting', self.name)
+        self.logger = logging.getLogger('Listener ' + self.name)
+        self.logger.info('start')
         self.trigger_timer = datetime.now()
 
         self.chat = twitch.Chat(
@@ -56,8 +62,9 @@ class StreamListener:
     def stop(self):
         """остановка процесса"""
         if getattr(self, 'process', None):
-            print('stopping process')
-            self.process.kill()
+            self.logger.info('stop')
+            if self.process.is_alive():
+                self.process.kill()
         else:
             self.chat.dispose()
             self.is_listening = False
@@ -82,7 +89,7 @@ class StreamListener:
             streamer.clips.append(clip)
             streamer.activity += 1
             session.commit()
-        print('starting', streamer.name, activity)
+        self.logger.info('saving buffer with activity: ' + str(activity))
 
         filename = f'{streamer.name}_{clip.id}.mp4'
         clip_path = self.clip_path + filename
@@ -99,11 +106,10 @@ class StreamListener:
     def _phrazes_handler(self, message):
         """обработчик сообщений чата"""
         text = message.lower()
-        print(self.name, text)
         self._chat_buffer_update()
         for phraze in self.phrazes:
             if phraze in text:
-                print(self.name, phraze, len(self.chat_buffer))
+                self.logger.info('trigger: ' + phraze)
                 self.chat_buffer.append(datetime.now())
                 break
         if len(self.chat_buffer) >= self.phraze_threshold and datetime.now() - self.trigger_timer > self.trigger_timeout:
@@ -117,13 +123,15 @@ class StreamListener:
         self.stream = self.session.streams(
             'https://www.twitch.tv/' + self.streamer.name)['best'].open()
         self.video = bytearray()
-        print('started listening stream', os.getpid())
+        self.logger.info('start listening stream')
         while True:
             if not self.is_listening:
                 break
-            try:
+            error_buffer = StringIO()
+            with redirect_stdout(error_buffer):
                 data = self.stream.read(self.recieving_bytes_amount)
-            except:
+            if 'unable to' in error_buffer.getvalue().lower():
+                self.logger.error('unable to reload')
                 sys.exit()
             try:
                 self.video += data
@@ -133,7 +141,7 @@ class StreamListener:
                 pass
     
     def _chat_listener(self):
-        print(self.name, 'starting listening chat')
+        self.logger.info('start listening chat')
         while True:
             messages = [message['message'] for message in self.chat.twitch_receive_messages()]
             for message in messages:
@@ -143,12 +151,12 @@ class StreamListener:
 
     def _save_by_timer(self, seconds):
         """функция для потока таймера запуска"""
-        print('starting count messages', seconds)
+        self.logger.info('start count activity ' + str(seconds))
         self.can_deleting_message_buffer = False
         time.sleep(seconds)
         phrazes_count = len(self.chat_buffer)
         self.can_deleting_message_buffer = True
-        print('starting wait timeout')
+        self.logger.info('starting wait timeout')
         time.sleep(self.save_timeout)
         self.save_buffer(phrazes_count)
         self.chat_buffer.clear()
@@ -170,6 +178,7 @@ class StreamListener:
     def load_phrazes(self):
         """загрузка фраз из базы данных"""
         session = self.session_maker()
+        session.add(self.streamer)
         self.name = self.streamer.name
         self.phraze_threshold = self.streamer.threshold
         phrazes = session.query(Trigger).all()
